# public/openapi.yaml
openapi: 3.0.3
info:
  title: ApplyCopilot REST API (Phase 1)
  description: API Contracts for Authentication and CV Import pipelines in ApplyCopilot V2.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
paths:
...
  /api/profile:
    get:
      summary: Retrieve complete profile
      description: Returns the user's basic data, experiences, education, projects, skills, references, active summaries, and parsed CV versions.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Complete profile retrieved successfully.
        '401':
          description: Unauthorized.
...
  /api/profile/experiences/{id}:
    put:
      summary: Update experience entry
      description: Updates the experience and reconciles its bullets.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Experience updated.
        '401':
          description: Unauthorized.
    delete:
      summary: Delete experience entry
      description: Deletes the experience and cascade deletes its bullets.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Experience deleted.
        '401':
          description: Unauthorized.
...
  /api/profile/skills:
    put:
      summary: Overwrite skills list
      description: Replaces the complete flat skills list for the profile.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
      responses:
        '200':
          description: Skills updated.
        '401':
          description: Unauthorized.
...
  /api/profile/references:
    put:
      summary: Overwrite references list
      description: Replaces the complete references list for the profile.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
      responses:
        '200':
          description: References updated.
        '401':
          description: Unauthorized.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: Session cookie generated by NextAuth.js